Refactor: split monolith to src modules
