+++ src/__init__.py
@@ -1,0 +1,3 @@
+from .api import AutoRewarderAPI
+
+__all__ = ["AutoRewarderAPI"]
