+++ src/history.py
@@ -1,0 +1,68 @@
+import os
+import json
+from datetime import datetime
+
+from .config import HISTORY_FILE_PATH, SETTINGS_FILE_PATH
+
+class HistoryManager:
+    def __init__(self):
+        self.history_file = HISTORY_FILE_PATH
+        self.settings_file = SETTINGS_FILE_PATH
+
+    def get_history(self):
+        if not os.path.exists(self.history_file) or os.path.getsize(self.history_file) == 0:
+            return []
+
+        try:
+            with open(self.history_file, "r", encoding="utf-8") as file:
+                history = json.load(file)
+
+                if not isinstance(history, list):
+                    raise ValueError("History data must be a list")
+
+                return history
+        except (json.JSONDecodeError, UnicodeDecodeError, ValueError):
+            self.log("[ERROR] History file was unreadable or damaged. Starting with a fresh one.")
+
+            # Keep the damaged history file as backup and start with a clean one.
+            backup_path = self.history_file + ".backup"
+
+            if os.path.exists(backup_path):
+                os.remove(backup_path)
+
+            os.replace(self.history_file, backup_path)
+
+            with open(self.history_file, "w", encoding="utf-8") as file:
+                json.dump([], file, indent=4)
+
+            return []
+
+    # Save search history to temp file first and then replace the original file to avoid data corruption
+    def save_history(self, history_list):
+        temp_file = self.history_file + ".tmp"
+
+        with open(temp_file, "w", encoding="utf-8") as file:
+            json.dump(history_list, file, indent=4)
+
+        # Replace the original file with the updated one
+        os.replace(temp_file, self.history_file)
+
+    # Add a search query to history JSON file
+    def add_to_history(self, query_text, status):
+        now = datetime.now()
+        current_date = now.strftime("%m-%d-%Y")
+        current_time = now.strftime("%H:%M:%S")
+
+        # Create a new record with date, time, query, and status
+        new_record = {
+            "date": current_date,
+            "time": current_time,
+            "query": query_text,
+            "status": status
+        }
+
+        history_list = self.get_history()
+
+        history_list.append(new_record)
+
+        self.save_history(history_list)
